various

+++ src/backend.zig
@@ -3,7 +3,7 @@ const root = @import("main.zig");
 pub const Host = @import("backend/host.zig");
 pub const OpenCl = @import("backend/opencl.zig");
 
-pub const Kind = enum(c_uint) {
+pub const Kind = enum(usize) {
     Host = 0,
     OpenCl = 1,
     // Cuda,
